Code moved from osn-output-signals to osn-output
